requirements.txt
beautifulsoup4==4.7.1
joblib==0.13.2
lxml==4.6.2
Markdown==3.0.1
pelican==4.0.1
PyYAML==5.1
requests==2.20.0
